Fix inclusion of changes to .idea files in commits
I realised that commits from contributors kept on including modifications to file(s) in the .idea directory in the project root, which is unsightly. It would appear that said file(s) are for individual configuration on different machines, and hence, the constant modifications across commits from different contributors. The .idea directory should be included in the .gitignore and I had used GitHub's default Android .gitignore during repo creation. I don't know if that file was inadequate or somebody messed this up afterwards, but I have instead replaced that .gitignore with a custom .gitignore that I always use in my Android projects which is tried and tested and gets the job done.