[WIP] Add map functionality
The map fragment has been added using the Google Maps Android API. At the moment, only the My Location functionality has been implemented. The API key is private, and should only be used locally in the debug build variant of our app and should *never* be added to the repo. Additionally, I implented an activity to display the Google Play Services attribution text, as is required by the attribution requirements of using the Google Maps Android API in our application.