Fixed a NullPointerException
calling getActivity() inside VideoFragment.initialize() threw a null pointer because the method as being called before the Fragment was attached to the Activity. Turns out the method call was unnecessary because the String being requested was already inside the current context. The method has been removed and the String used instead