Add log in help dialog
I want to make elements in the message actionable, like phone numbers and links that will launch some app or activity that can perform the desired action. Additionally, I'm pretty sure I need to write code to alter the visibility of the new TextView widget as logging in is attempted and a progress bar is shown.