Update appearance of recycler view lists
I removed the padding above and below the recycler views because I think the appearance of the cut off seems weird, definitely at the end of the list. As for the padding above, I feel that the content should meet at the toolbar's lower seam as we scroll down the list. I also removed scrollbars because they just don't seem necessary, especially for the News section. In production, the number of news items would be indeterminate, so I don't see a need for the scroll bar there. As for the information section, although the list is finite, I still feel the scrollbar is unnecessary, as well as disctracting.