Refactor UI from run time to static
Since the cards of the InfoFragment are really just for navigational purposes (to go to other activities), I didn't see a reason for the use of a RecyclerView and corresponding Adapter for it. These cards are not dynamic and I think this is simpler and doesn't incur the run time overhead of the old way.