Improve appeareance of Nav Drawer
My primary goal was for the Nav Drawer to be drawn behind the status bar on API v19+. However, I ran into the white status bar issue in the SASLoginActivity and SettingsActivity so I had to make some additional adjustments. Additionally, I replaced the log out icon and removed the section header name for Settings in the drawer.