Colorise Info Activities
For API v21+ the toolbar and status bar colours of the Info Activities are now derived from the same Material colour palette as the background colour of the images in their corresponding cards in the Info Fragment. At the moment, some unsightly clipping is happening at the top of the videos listing. The layout of the videos listing is going to be changed soon so it will be fixed then. Additionally, I want to bring this over to API v19 to take advantage of translucent status bars. Also, the colour values may need some tweaking. Finally, I'm pretty certain that there is a cleaner way to implement this.