Prevent memory leaks with AsyncTask
I made the AsyncTask class static so that it won't hold the implicit reference to the Activity containing it so that it won't prevent garbage collection on Activity destruction. Additionally, I use WeakReferences to hold references to members of the Activity for the same reason. The use of WeakReferences facilitates asynchronously updating Views in the Activity on the UI thread in onPostExecute() safely. The code seems unpleasant and can probably be refactored.

This same thing will also have to be done for the other AsyncTasks being used elsewhere in the project.