Fixed #4 - Video player Activity crash
We are using VectorDrawableCompat in the app, so pre-Lollipop API versions cannot directly use VectorDrawables, so the srcCompat attribute needs to be used in XML and not the src one, or they will crash. VectorDrawables are not supported in compound views at this time by the Support Library. There seems to be a way to do it programatically using StateListDrawables, but that seems unnecessary.

Fixes #4