Fix NewsDetailActivity transition crash
When tapping on a news Card, the app would crash due to some error with a TextView in the NewsDetailActivity. I'm not sure when this bug was introduced, perhaps it was in my previous commit, dae6b1b983a0c37e15d4864f99229ff3cb5b9acd? I manually undid the changes introduced in 39239cdf75470dbc052d780be890d1936549884b and related commit 43df9939365bcc8421cc2da1526664dd3961bf11 to resolve this issue. Additionally, the animation for the transition introduced in the former commit was not smooth for the shared news title TextViews, so this feature enhancement was still a WIP anyway.