Fix cutoff of Views in Login Activity
Instead of hardcoding the height of the ImageView, I now use LinearLayout weighting to make it take up 2/3 of the screen height in relation to the rest of the buttons on the screen. This should fix the issue of some of the buttons being cutoff from the screen on devices with smaller screens. I also removed the redundant LinearLayout wrapping the ImageView.